created DeeplTranslator,superclass of translate text and file DAO
